+++ services/dask-sidecar/src/simcore_service_dask_sidecar/utils/files/__init__.py
@@ -1,0 +1,25 @@
+from ._copy import CHUNK_SIZE
+from ._download import check_need_unzipping, pull_file_from_remote
+from ._progress import LogPublishingCB
+from ._s3 import (
+    HTTP_FILE_SYSTEM_SCHEMES,
+    S3_FILE_SYSTEM_SCHEMES,
+    ClientKWArgsDict,
+    S3FsSettingsDict,
+    _s3fs_settings_from_s3_settings,
+)
+from ._upload import MIMETYPE_APPLICATION_ZIP, push_file_to_remote
+
+__all__: tuple[str, ...] = (
+    "CHUNK_SIZE",
+    "HTTP_FILE_SYSTEM_SCHEMES",
+    "MIMETYPE_APPLICATION_ZIP",
+    "S3_FILE_SYSTEM_SCHEMES",
+    "ClientKWArgsDict",
+    "LogPublishingCB",
+    "S3FsSettingsDict",
+    "_s3fs_settings_from_s3_settings",
+    "check_need_unzipping",
+    "pull_file_from_remote",
+    "push_file_to_remote",
+)
